Move GridLineNames and its iterator into types/named.rs
